+++ iowmongotools/cluster.py
@@ -1,0 +1,8 @@
+""" Manages mongo cluster """
+
+
+class Cluster(object):
+    """ Represents mongo cluster """
+
+    def __init__(self):
+        pass